Add silence to audio for better results
